// src/types/payload.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Standard Webhook Payload
 */
export interface Payload {
  type: string;
  timestamp: string;
  data: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}